+++ src/main/java/um/tesoreria/core/service/ArancelTipoService.java
@@ -7,6 +7,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import jakarta.annotation.Resource;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
@@ -23,16 +24,18 @@ import lombok.extern.slf4j.Slf4j;
  */
 @Service
 @Slf4j
+@RequiredArgsConstructor
 public class ArancelTipoService {
 
-	@Resource
-	private ArancelTipoRepository repository;
-
-	@Resource
-	private ArancelTipoLectivoRepository arancelTipoLectivoRepository;
+	private final ArancelTipoRepository repository;
+	private final ArancelTipoLectivoRepository arancelTipoLectivoRepository;
 
 	public List<ArancelTipo> findAll() {
 		return repository.findAll(Sort.by("arancelTipoId").ascending());
+	}
+
+	public List<ArancelTipo> findAllHabilitados() {
+		return repository.findAllByHabilitado((byte) 1, Sort.by("arancelTipoId").ascending());
 	}
 
 	public List<ArancelTipoLectivo> findAllByLectivoId(Integer lectivoId) {
@@ -60,16 +63,19 @@ public class ArancelTipoService {
 
 	public ArancelTipo add(ArancelTipo arancelTipo) {
 		arancelTipo = repository.save(arancelTipo);
-		log.debug("ArancelTipo -> " + arancelTipo);
+		log.debug("ArancelTipo added -> " + arancelTipo.jsonify());
 		return arancelTipo;
 	}
 
 	public ArancelTipo update(ArancelTipo newArancelTipo, Integer arancelTipoId) {
+		log.debug("ArancelTipo new -> {}", newArancelTipo.jsonify());
 		return repository.findByArancelTipoId(arancelTipoId).map(arancelTipo -> {
-			arancelTipo = new ArancelTipo(newArancelTipo.getArancelTipoId(), newArancelTipo.getDescripcion(),
-					newArancelTipo.getMedioArancel(), newArancelTipo.getArancelTipoIdCompleto());
+			arancelTipo.setDescripcion(newArancelTipo.getDescripcion());
+			arancelTipo.setMedioArancel(newArancelTipo.getMedioArancel());
+			arancelTipo.setArancelTipoIdCompleto(newArancelTipo.getArancelTipoIdCompleto());
+			arancelTipo.setHabilitado(newArancelTipo.getHabilitado());
 			arancelTipo = repository.save(arancelTipo);
-			log.debug("ArancelTipo -> " + arancelTipo);
+			log.debug("ArancelTipo updated -> " + arancelTipo.jsonify());
 			return arancelTipo;
 		}).orElseThrow(() -> new ArancelTipoException(arancelTipoId));
 	}